Automatically install the default serializers module
This will make it easier to create a default value for the serializerModule
parameter.
